Sửa texture

+++ Pipe.h
@@ -13,7 +13,7 @@ protected:
 	BOOLEAN isGoInside;
 public:
 	CPipe(float x, float y,
-		float cell_width, float cell_height, int height, int spriteId_top_left, int spriteId_top_right, int spriteId_bot_left, int spriteId_bot_right, int isGoInside) :CGameObject(x, y)
+		float cell_width, float cell_height, float height, int spriteId_top_left, int spriteId_top_right, int spriteId_bot_left, int spriteId_bot_right, int isGoInside) :CGameObject(x, y)
 	{
 		this->height = height;
 		this->cellWidth = cell_width;
